backend/requirements.txt

fastapi
uvicorn
pydantic
requests
pandas
firebase-admin
python-dotenv
